Disable code coverage profiler after tests
Summary:
Enabling the code coverage profiler has global side-effects, which can
affect subsequent tests running in the same binary. For instance, the
profiler retains strong references to Domains, which can in turn affect
what gets garbage collected.

Ensure that it is disabled when the code coverage tests are torn down.

Reviewed By: avp

Differential Revision: D60212750

fbshipit-source-id: 4746952d184d7a1ef2bea2a9ab288ceff01e4741