Add stack promotion for scope instances
Summary:
Extend SimpleStackPromotion to identify entire scope instances that
never escape and convert them to stack locations. This allows
variables that escape in general to still be stack promoted, if they
are in a scope that never escapes.

This allows scopes instances that become non-escaping (e.g. due to
inlining) to be removed. Perhaps more importantly, they allow the init
scope emitted for scoped-for-loops to be stack promoted independently
of the scope in the body.

Reviewed By: avp

Differential Revision: D60474165

fbshipit-source-id: 1f912127f661d75d4445ead8d6b38b93c7f44a8c