Reenable blocking on OG collection in a YG collection
Summary:
In the new WeakMap implementation, entry slots are stored separately in
a ManagedChunkedList. In `markWeakRoots()`, we only need to get mark
bits of the owner maps and key objects, instead of accessing heap
objects, so inconsistent YG won't affect the correctness. It's safe to
reenable blocking on OG collection in a YG collection.

Reviewed By: neildhar

Differential Revision: D60475177

fbshipit-source-id: a2b6c81c41d845389a72c25f4629d3a1fc5a2ec5