Make TagType a little narrower.
Summary:
We got in this code in HermesValue.h:
```
  inline static constexpr unsigned combineETags(ETag aTag, ETag bTag) {
    using UTagType = std::make_unsigned<TagType>::type;
    auto a = static_cast<UTagType>(aTag), b = static_cast<UTagType>(bTag);
    return ((a & kETagMask) << kETagWidth) | (b & kETagMask);
  }
```
Line 139 is the return statement:
```
xplat/static_h/include/hermes/VM/HermesValue.h:139:44: error: implicit conversion loses integer precision: 'UTagType' (aka 'unsigned long') to 'unsigned int' [-Werror,-Wshorten-64-to-32]
    return ((a & kETagMask) << kETagWidth) | (b & kETagMask);
    ~~~~~~ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~
```

TagType is defined as the HVTagType in sh_legacy_value.h.  We're returning unsigned here; the simplest thing seemed to be to make the width of HVTagType match that of unsigned.  It doesn't seem to cause any problems, and it fixes the error.

Reviewed By: neildhar

Differential Revision: D56072293

fbshipit-source-id: fca42484f6ecbfd108115221e5d8d740c0c07e38