Check for overflow in RegexBytecode.
Summary:
We were getting narrowing errors in RegexBytecode.h.
```
xplat/static_h/include/hermes/Regex/RegexBytecode.h:340:26: error: implicit conversion loses integer precision: 'unsigned long' to 'uint32_t' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]
    return bytes_.size() - sizeof(RegexBytecodeHeader);
    ~~~~~~ ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
The code in question is:
```
  /// \return the current offset in the stream, which is where the next
  /// instruction will be emitted. Note the header is omitted.
  uint32_t currentOffset() const {
    return bytes_.size() - sizeof(RegexBytecodeHeader);
  }
```

This is a real potential overflow; nothing (AFAICT) prevents us from making a bytecode file more than 4GB in size.  If we did, this offsets and jump values would be wrong.

I propose that we solve this with overflow checking as we add bytecodes.  I added a sizeCheck operation in HermesSafeMath: ensure that a new value in a wide type will not overflow a possibly narrower type, calling hermes_fatal if it does.  I added this to all the emit operations; this now justifies the use of an unsafeNarrow in currentOffset.

Reviewed By: tmikov

Differential Revision: D56072749

fbshipit-source-id: 3cc7ae94cea6554681060fb76a83d0179cf216fc