Fix narrowing errors in BitArray.h.
Summary:
There were integer precision errors in BitArray.h.  This problems were with remainder (%) operations, whose results were cast from size_t to a narrower type (unsigned).  This was justified by the divisor being small enough that the remainder (< the divisor) fits in the narrower type.

The divisor was always constexpr, so I made a

Reviewed By: neildhar

Differential Revision: D56084289

fbshipit-source-id: f7930da3b727626724cd593148dedaa5922dae23