Remove usage of CompareBranchInst outside lowering
Summary:
We currently do not handle `CompareBranchInst` in `InstSimplify` and
`SimplifyCFG`, which means that we may miss out on optimisations for
them. To avoid duplicating optimisations implemented for
`BinaryOperatorInst` and `CondBranchInst`, don't create this instruction
until lowering.

While there currently does not appear to be any performance sensitive
usage of this before lowering, this allows us to simplify thinking about
where this instruction might actually be present.

Reviewed By: avp

Differential Revision: D60996879

fbshipit-source-id: c1878ca1480534cab047d5725a3ec0cc3e05734f