Handle dead functions in TypeInference by resetting Variables
Summary:
It was possible for an enclosing function to be removed without
its child functions being removed, resulting in their attempting to load
from a `VariableScope` that was never created and never stored to.

To fix this, reset "NoType" variables to their prepass state,
much like we do for function params/return types.

Reviewed By: neildhar

Differential Revision: D60801803

fbshipit-source-id: 6762dec237aec46468b4aa6868f02b458fe804d3