Run until valid pause location on step-out
Summary:
Stepping out should also ensure that there's a valid place to continue
stepping from.

Now that there will be instructions without locs, we must not pause on
those. We have to always continue the step to a valid stopping point.
It's possible that StepOut should've always done this, but we used to
always have locations for every instruction.

This does cause changes to tests because we no longer
stop on a statement=0 instruction on StepOut, making the behavior
consistent across stepping modes.

Reviewed By: dannysu

Differential Revision: D59247325

fbshipit-source-id: 42b011f0972c4e9d471d5fbd1628a3eb127cc3a5