Add calleeIsAlwaysClosure flag to calls
Summary:
Add a flag to calls indicating whether the callee is always known to be
a closure. This is used to signal that while the target on the call is
the only possible function that can be invoked at this call, there may
be other non-closure values at a call, which would throw. Inlining can
use this information to determine whether a runtime type check is needed
when inlining a call.

Reviewed By: avp

Differential Revision: D60950543

fbshipit-source-id: be40c931dd678e58f338cb1b6cc154e46ae7ef74