Inline functions stored to TDZ/IDZ variables
Summary:
Extend FunctionAnalysis to trace closures stored to variables where the
only other store is Empty/Uninit. This exploits the fact that there
must be checks before those values can be passed to a call, so we are
always guaranteed that a call loading from such a variable will not
have Empty or Uninit.

This is the first step towards extending this approach to all other
non-closure values that are stored to variable. As long as we know we
have the only store that is potentially a closure, we can assume that
the target of the call will be that closure, with a dynamic check on
the type to ensure safety.

Reviewed By: davedets

Differential Revision: D60487690

fbshipit-source-id: e0d2233b324a7da8b1db137e0fcf9f87850bcdba