Add path provided from "-L" flag into rpath
Summary:
For path in SHERMES_CC_LIB_PATH, we add corresponding "-L" flag and
"-Wl,-rpath" flag to CC. But for path provided in "-L" to shermes, we
only add "-L" to CC. This is kind of inconsistent (though the shermes
flag and CC flag are both "-L").

This diff also adds path in "-L" flag of shermes to rpath. It's the
easiest solution to unblock following diffs. It can be changed once we
have a conclusion on the shermes CLI flags.

Reviewed By: neildhar

Differential Revision: D61742563

fbshipit-source-id: 572dd94659ea87f1ca8768bdca30b9e668cb1cfa