Handle printing empty ASTs with docblocks
Summary:
Original Author: pieterv@meta.com
Original Git: eda982396ea3381c69a2fde5d508f22ccfab0caa
Original Reviewed By: gkz
Original Revision: D60191972

If a program body has no statements, there is no need to use prettier to print it. Instead just return the docblock comment or an empty string.

This is needed due to an issue with prettier printing empty files with comments. With the hermes-transform infra today we perform comment attachment during "parse" so that comments are attached when transforming the AST. Typically however with prettier the attachment is done during printing, but this can be skipped by removing the `comments` array off the program node (so there are not comments to attach). This works great for normal nodes as it prints based off each nodes `comments` array. But for the docblock if there is no statement in the body to attach it to we need to add it to the program comments array but this triggers attachment which fails with our AST format.

Reviewed By: avp

Differential Revision: D61574212

fbshipit-source-id: bcd77b7f4bc19ec6ada8129c1659a773424c7089