CMake: prepare build infra for the JIT
Summary:
A new CMake option `HERMESVM_ALLOW_JIT`. Value 0 disabled it, value 1
enables it on platforms where it is supported, value 2 forces it to be
enabled.

This is communicated to code by passing a define with the values 1 or 2
(when 0 the define is not defined at all).

`hermes/VM/JIT/Config.h` checks various conditions and defines
`HERMESVM_JIT` as either 0 or 1.

Reviewed By: avp

Differential Revision: D61752807

fbshipit-source-id: 5834d8119bc6f8015f01995948fe8974098767d5