Fix double to int32 conversion for negative numbers
Summary:
When checking for whether a double can be an int32, we currently cast
to uint64. This ends up using `fcvtzu` on arm64, which means that
negative numbers get turned into 0 and fail the fast path check.

Instead, we should cast through `int64`, which will do a signed
conversion. We still need the `uint64` since left shift of negative
numbers is UB until C++20.

Reviewed By: tmikov

Differential Revision: D62084829

fbshipit-source-id: 9a36a6d36e0918b423bc8c9c59afd9ded803c15d