Refactor captured Variables into CapturedState
Summary:
These 3 captured variables are always used together, so we should just
bundle them together in a struct. We will also be adding more captured
variables in the classes stack, so this makes it easier and less error
prone to ensure the correct propagation of new `Variable *`s that we
add.

Reviewed By: neildhar

Differential Revision: D61566046

fbshipit-source-id: 79b3923221f3c2fc8093dfff211ca654a4dfc2f7