Disable -fsanitize=function on Apple platforms
Summary:
`-fsanitize=function` relies on storing data before a function pointer
that describes the type of the function. However, on Apple platforms,
this data gets treated as part of the preceding function and can be
moved around or deleted. This results in UBSan false positives.

To address this, pass `-fno-sanitize=function` when compiling for Apple
platforms with UBSan enabled.

See https://github.com/llvm/llvm-project/issues/109074

Reviewed By: fbmal7

Differential Revision: D62921741

fbshipit-source-id: da7fe78fc09840c90a48e6dfcf32ed14c514c323