Optimise FastArray store bounds check
Summary:
The store bounds check for FastArray currently converts the length from
a double to an int and then performs an integer comparison against the
index. Instead, directly retrieve the size from the ArrayStorage, as we
do for loads.

This also avoids needing to unbox the length property when HV32 is on.

Reviewed By: tmikov

Differential Revision: D62581590

fbshipit-source-id: 81c93e4c1897d6d17f76cf198de9ef08a0c4fdd5