Properly enable compiler statistics.
Summary:
We have a number of uses of the Statistic class in the compiler, but
there was no way to enable their gathering and printing.  This diff
fixes that.

There were a couple of problems, but with the changes here, the
workflow described in the test plan gets the statistics.

Reviewed By: avp

Differential Revision: D62511937

fbshipit-source-id: 359b858e697b56013844f941ab6a55cdbe04d6d5