implement double to integer truncation without UB
Summary:
On the platforms we support, implement double truncation either with the
available intrinsics or with inline assembly.

All the combinations of supported platforms and compilers have been
compiled on Godbolt, with output manually inspected, but not all have
run tests, since we don't have CI setup for all of them.

Reviewed By: avp

Differential Revision: D63240632

fbshipit-source-id: 7f43dba292ab8929a29da872fe2d3d3c238b8d9e