Update CreateThis IR & Bytecode instructions for new `this` allocation strategy
Summary:
Now that `NativeConstructor`s make their own `this`, we should update
the interpreter to not make an object if it's going to call a
NativeConstructor. It is not implemented yet, but ES6 classes will
also make their own `this`, so we can also safely skip object creation
for them.

`createThis` will now sometimes discard the input
`new.target.prototype`. Rather than have this wasted work, we fuse the
`.prototype` read into the createThis instruction. This means that
`createThis` must also perform all the same caching logic that normal
property reads do.

In the event of a `new foo()` expression, we know that `new.target` will
be set to `foo`, so we create a specialized `CreateThisForNew` which
only takes in the callee.

The native backend for now calls out to `createThisForConstruct_RJS`,
but is made faster next diff.

Reviewed By: neildhar

Differential Revision: D59017738

fbshipit-source-id: d88d64b03addea19bc1e3113f91abda40ebbeca7