Annotate which Callable CellKinds make their own `this`
Summary:
Create new CellKind ranges to describe the different kinds of Callables,
with respect to `this` allocation during a construct call. Reshuffle the
Callable CellKinds to sit in these new regions. This change is somewhat
complicated by the following 2 points:

1. NativeFunction is an existing range over [NativeFunction,
NativeConstructor, JSCallableProxy, FinalizableNativeFunction].
Therefore, we need to ensure that in this re-ordering, these CellKinds
remain contiguous. Luckily, these native functions only span 2 different
constructor regions, so this is possible.
2. We need to make sure to move BigIntPrimitive and DummyObject up
before JSObject, since the Object range is defined to be [JSObject,
JSFunction].

Reviewed By: avp

Differential Revision: D61165947

fbshipit-source-id: af4b1bf61e7574d8ea4c33c9116dad7bd3c5f48c