Return past breakpoints on a call
Summary:
The interpreter inspects the IP up the stack to determine where to
return to, so check whether there's a breakpoint there before doing
`nextInstCall`. The check is now faster because we maintain a count of
how many installed breakpoints each `CodeBlock` has when the debugger is
enabled and check that before doing the more expensive map lookup on
every Ret.

Reviewed By: dannysu

Differential Revision: D63920292

fbshipit-source-id: f3196ebbde59cc49b53c2415a97e3e4728a8291d