Use getRealOpCode instead of getNextOffset in debugger
Summary:
Step-out calculation must consider that there's a already a
breakpoint up the stack, so it has to check and use the real opcode
when calculating where to set the new step-out breakpoint.

`getNextOffset` is a function that shouldn't exist because it only is
used in the debugger when we should be accounting for breakpoints.

Reviewed By: dannysu

Differential Revision: D63863793

fbshipit-source-id: 8562d96c24c2b1cd50d91b12ff438b34bb33b127