JIT: Don't output code in the Emitter constructor
Summary:
The `Emitter` constructor was allocating registers and setting up the
stack frame, leading to confusing output when dumping code.
The "JIT compilation of FunctionID..." line and the name of the function
would come after a bunch of code had already been emitted to set up the
stack frame.

Split the code into an `enter()` function that we call before compiling
all the blocks, to make sure any actual work is deferred until `enter()`
is called, avoiding work in the constructor.

Makes the dump much clearer to read.

Reviewed By: neildhar

Differential Revision: D64061957

fbshipit-source-id: a718e278c2ab660334b4b563b675dd546c6ddc59