Reject invalid `super` property properties
Summary:
Implement proper validation for `super` references in SemanticResolver.
The core idea is that all functions defined with method syntax can use
`super`, and all class field initializers can use `super`. Arrow
function also take the `super` binding of where they are defined.

Because the scope of this super binding does not correspond 1:1 with any
particular function, we create a new field `hasSuperBinding` directly on
`SemanticResolver`. All non-arrow functions will update this field on
entry. The decision to set or clear the flag is made by inspecting the
parent.

Reviewed By: avp

Differential Revision: D62451111

fbshipit-source-id: dcec7a5db5970fe4111718d96859a6a68ce70206