Support `this` and `new.target` when debugging arrow functions
Summary:
We want to support using `this` and `new.target` in the debugger. In
order to do this, there are 2 situations we deal with separately:
stopped in an arrow function, or stopped in a non-arrow function. The
expression that is given to the debugger, e.g. `this.a`, is wrapped into
its own function. When stopped in a non-arrow, we can generate the IR
for the expression as if it was in a normal function. This means that
for `this`, it will generate a `LoadParameter` for index 0. Then, when
we call this function containing the expression, we can pass it the
parameters for `new` and `this`, so that the IR instructions produce the
expected value. Arrow functions have to be treated differently.

For arrow functions, we can't copy the `this`/ `new.target` of the
function we are stopped in, we must use the captured state variables.
So, when we create the `EvalCompilationDataInst` for an arrow function,
we must save the values of the captured state into it. Then, when
restarting compilation for the eval expression, if we see that
`EvalCompilationDataInst` has non-null captured state, then we set up
the function containing the expression as an arrow function. We also
copy the `Variable*`s into the function context of this new arrow
function. That way, when we encounter `this`, we don't generate a
`LoadParam`, but rather we generate a load to the captured variable.

Reviewed By: avp

Differential Revision: D63131974

fbshipit-source-id: fbf62b7127524cd8b83ffc39c72671692e9c64c0