Fix CodeMotion sinking for irreducible loops
Summary:
Sinking in CodeMotion is invalid (and suboptimal) if an instruction is
being sunk into a different loop. We determine whether the loops are
the same by comparing loop heaaders, but this does not handle the case
where the header for a loop cannot be determined (and is therefore
null).

Add a null check to handle this case.

Reviewed By: tmikov

Differential Revision: D61643578

fbshipit-source-id: 6922235e66ba6dc6ce30d2a3bf25ac85df4f26f5