Simplify throwing exception in parent for ProhibitInvoke
Summary:
The logic for throwing the ProhibitInvoke exception is currently
complicated by the fact that it should be thrown from the caller even
though the check technically happens in the callee. This means that some
state needs to be manipulated after the exception is thrown to ensure
that the resulting stack trace is correct.

Simplify this by performing the check at the Call instruction when doing
a tail call, so that no state needs to be manipulated. In the case of a
recursive invocation of the interpreter, perform a check right at the
entry of the interpreter before any of the state has been set up.

This is useful in the next diff, which changes our stack trace
generation to rely on the saved IP instead of deferring to the
interpreter. Now that the exception is actually thrown at the Call, the
saved IP will automatically be correct.

Reviewed By: fbmal7

Differential Revision: D64159274

fbshipit-source-id: 31a81b1e4aef435a77a46c8a4a338a1d0087174a