Use IP in runtime for top frame in recordStackTrace
Summary:
We currently skip recording a stack trace when the top frame is in the
interpreter in favour of letting the Interpreter call back into
recordStackTrace with the IP and CodeBlock.

Change it so we use the IP from the runtime and retrieve the CodeBlock
from the call frame instead. This seems simpler and more consistent.

Importantly, this simplifies producing stack traces in the JIT, because
we do not need to recreate the Interpreter functionality.

Reviewed By: fbmal7

Differential Revision: D64159275

fbshipit-source-id: b9b145c78e5050fd6cef94939b6bde14f88d206c