Add call handler for the JIT
Summary:
Add a new JIT-specific call handler that also writes the current IP to
the SavedIP slot in the outgoing registers. This is necessary for us to
produce stack traces for call stacks with JITted functions in them.

In upcoming diffs, we can move the population of the call stack to the
code emitted by the JIT, which will allow us to do the fast path check
and call to other JITted functions without the indirection.

Reviewed By: avp

Differential Revision: D64475475

fbshipit-source-id: f9340783dd74a02575cb2ab355f624b1f498f382