Add register classes to HBC backend
Summary:
Much like in the native backend, the HBC backend now has a class/index
pair in each `Register`. There's type-specific register allocation which
ensures that registers which are number or non-pointer don't interfere
with each other or get clobbered by possibly-pointer registers.

Most of the diff is dealing with this in calls/spilling and throughout
the post-regalloc pipeline which relies on it.

There's new functions in `HVMRegisterAllocator` for converting to HVM
registers from class-separated `Register`, because the VM still just has
a single list of registers.

Spilling is handled by converting all type-specific registers into
general-purpose registers and spilling as usual; this seemed like the
easiest way to do it, and it avoids adding more logic directly in
the register allocator for something that doesn't happen very often.

There's some minor changes required in figuring out what the "maximum"
register is, given that there's multiple classes now and the stack frame
still needs to use them for calls.

Now that the registers have classes, we also update the dumper to dump
the classes and avoid printing a destination for instructions without
output.

Reviewed By: tmikov

Differential Revision: D62992428

fbshipit-source-id: 81437d95bd98b17f870c1bd17ec47159ab257599