Minor optimisations to _callWrapper
Summary:
Simplify callWrapper so that it does not need to keep the IP and
currentFrame across the call. Also add a LIKELY annotation on the try.

I do not expect this to really affect performance, but I was curious to
understand the effect on the emitted code (which does improve). For the
IP in particular, we may want to do something similar to avoid IPSaver
in the interpreter (as the TODO there suggests).

Reviewed By: avp

Differential Revision: D65010242

fbshipit-source-id: 5cbef4a248497cf892fa84f3bc64e775f6e50f1b