Simplify inverted conditional jumps
Summary:
Currently, inverted conditional jumps incur a taken branch in both the
taken and not-taken cases. Use the negated condition code instead so
we do not take any branches in the not-taken case.

Note that it is possible to further improve the taken case when the
operands are not known to be numbers by using a third set of condition
codes which are like the inverted ones but produce false for NaN. If we
do that, we can do the taken jump before the NaN check, knowing that the
NaN check afterwards will take us to the slow path. However, this will
diverge from the known-number case, since that does not have a NaN check
and so will need to use the form in this diff.

Reviewed By: tmikov

Differential Revision: D65185512

fbshipit-source-id: a7316d576f49a3f01c0687e5f7062c231744432a